@@ -1,7 +1,7 @@
 {
   "name": "quickapp",
-  "version": "2.5.3",
-  "description": "Angular5 (with ASP.NET Core) startup project template with complete login, user and role management. Plus other useful services for Quick Application Development",
+  "version": "2.6.0",
+  "description": "Angular6 (with ASP.NET Core 2.1) startup project template with complete login, user and role management. Plus other useful services for Quick Application Development",
   "author": {
     "name": "Ebenezer Monney",
     "email": "info@ebenmonney.com",
@@ -12,66 +12,64 @@
   "repository": "https://github.com/emonney/quickapp.git",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --extract-css",
+    "start": "ng serve",
     "build": "ng build --extract-css",
-    "build:ssr": "npm run build -- --app=ssr --output-hashing=media",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "dependencies": {
-    "@angular/animations": "^5.2.5",
-    "@angular/common": "^5.2.5",
-    "@angular/compiler": "^5.2.5",
-    "@angular/core": "^5.2.5",
-    "@angular/forms": "^5.2.5",
-    "@angular/http": "^5.2.5",
-    "@angular/platform-browser": "^5.2.5",
-    "@angular/platform-browser-dynamic": "^5.2.5",
-    "@angular/router": "^5.2.5",
-    "@ngx-translate/core": "^9.1.1",
-    "@swimlane/ngx-datatable": "^11.2.0",
+    "@angular/animations": "^6.0.7",
+    "@angular/common": "^6.0.7",
+    "@angular/compiler": "^6.0.7",
+    "@angular/core": "^6.0.7",
+    "@angular/forms": "^6.0.7",
+    "@angular/platform-browser": "^6.0.7",
+    "@angular/platform-browser-dynamic": "^6.0.7",
+    "@angular/router": "^6.0.7",
+    "@ngx-translate/core": "^10.0.2",
+    "@swimlane/ngx-datatable": "^13.0.1",
     "bootstrap": "^3.3.7",
-    "bootstrap-datepicker": "^1.7.1",
-    "bootstrap-select": "^1.12.4",
+    "bootstrap-datepicker": "^1.8.0",
+    "bootstrap-select": "^1.13.1",
     "bootstrap-toggle": "^2.2.2",
-    "chart.js": "^2.7.1",
-    "core-js": "^2.5.3",
+    "chart.js": "^2.7.2",
+    "core-js": "^2.5.7",
     "font-awesome": "^4.7.0",
     "jquery": "^3.3.1",
     "ng2-charts": "^1.6.0",
-    "ng2-toasty": "^4.0.3",
-    "ngx-bootstrap": "^2.0.2",
-    "rxjs": "^5.5.6",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.20"
+    "ngx-bootstrap": "^3.0.1",
+    "ngx-toasta": "^0.1.0",
+    "rxjs": "^6.2.1",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.0",
-    "@angular/compiler-cli": "^5.2.5",
-    "@angular/language-service": "^5.2.5",
-    "@types/jasmine": "~2.8.6",
+    "@angular/compiler-cli": "^6.0.7",
+    "@angular-devkit/build-angular": "~0.6.8",
+    "typescript": "~2.7.2",
+    "@angular/cli": "~6.0.8",
+    "@angular/language-service": "^6.0.7",
+    "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/jquery": "^3.3.0",
-    "@types/node": "~9.4.6",
-    "codelyzer": "^4.1.0",
+    "@types/jquery": "^3.3.4",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.2.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~1.7.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.4.1",
-    "karma-jasmine": "~1.1.1",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.0",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "protractor": "~5.3.2",
+    "ts-node": "~5.0.1",
+    "tslint": "~5.9.1"
   },
   "keywords": [
     "ng",
     "angular",
     "angular2",
-    "angular5",
+    "angular6",
     "angularX",
     "aspnetcore",
     "aspnet-core",
@@ -82,7 +80,7 @@
     "quickapplication",
     "rad",
     "ng2",
-    "ng5",
+    "ng6",
     "ngX"
   ]
 }
