@@ -10,7 +10,7 @@ export class AppPage {
     return browser.get('/');
   }
 
-  getParagraphText() {
-    return element(by.css('app-root h1')).getText();
+  getAppTitle() {
+    return element(by.css('app-root .appTitle')).getText();
   }
 }
