@@ -4,7 +4,7 @@
 // ====================================================
 
 import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
+import { Routes, RouterModule } from '@angular/router';
 
 import { LoginComponent } from "./components/login/login.component";
 import { HomeComponent } from "./components/home/home.component";
@@ -18,26 +18,22 @@ import { AuthService } from './services/auth.service';
 import { AuthGuard } from './services/auth-guard.service';
 
 
+const routes: Routes = [
+  { path: "", component: HomeComponent, canActivate: [AuthGuard], data: { title: "Home" } },
+  { path: "login", component: LoginComponent, data: { title: "Login" } },
+  { path: "customers", component: CustomersComponent, canActivate: [AuthGuard], data: { title: "Customers" } },
+  { path: "products", component: ProductsComponent, canActivate: [AuthGuard], data: { title: "Products" } },
+  { path: "orders", component: OrdersComponent, canActivate: [AuthGuard], data: { title: "Orders" } },
+  { path: "settings", component: SettingsComponent, canActivate: [AuthGuard], data: { title: "Settings" } },
+  { path: "about", component: AboutComponent, data: { title: "About Us" } },
+  { path: "home", redirectTo: "/", pathMatch: "full" },
+  { path: "**", component: NotFoundComponent, data: { title: "Page Not Found" } }
+];
+
 
 @NgModule({
-    imports: [
-        RouterModule.forRoot([
-            { path: "", component: HomeComponent, canActivate: [AuthGuard], data: { title: "Home" } },
-            { path: "login", component: LoginComponent, data: { title: "Login" } },
-            { path: "customers", component: CustomersComponent, canActivate: [AuthGuard], data: { title: "Customers" } },
-            { path: "products", component: ProductsComponent, canActivate: [AuthGuard], data: { title: "Products" } },
-            { path: "orders", component: OrdersComponent, canActivate: [AuthGuard], data: { title: "Orders" } },
-            { path: "settings", component: SettingsComponent, canActivate: [AuthGuard], data: { title: "Settings" } },
-            { path: "about", component: AboutComponent, data: { title: "About Us" } },
-            { path: "home", redirectTo: "/", pathMatch: "full" },
-            { path: "**", component: NotFoundComponent, data: { title: "Page Not Found" } },
-        ])
-    ],
-    exports: [
-        RouterModule
-    ],
-    providers: [
-        AuthService, AuthGuard
-    ]
+  imports: [RouterModule.forRoot(routes)],
+  exports: [RouterModule],
+  providers: [AuthService, AuthGuard]
 })
 export class AppRoutingModule { }
