@@ -12,7 +12,7 @@ import { HttpClientModule } from '@angular/common/http';
 
 import { TranslateModule, TranslateLoader } from "@ngx-translate/core";
 import { NgxDatatableModule } from '@swimlane/ngx-datatable';
-import { ToastyModule } from 'ng2-toasty';
+import { ToastaModule } from 'ngx-toasta';
 import { ModalModule } from 'ngx-bootstrap/modal';
 import { TooltipModule } from "ngx-bootstrap/tooltip";
 import { PopoverModule } from "ngx-bootstrap/popover";
@@ -67,66 +67,66 @@ import { RoleEditorComponent } from "./components/controls/role-editor.component
 
 
 @NgModule({
-    imports: [
-        BrowserModule,
-        BrowserAnimationsModule,
-        HttpClientModule,
-        FormsModule,
-        AppRoutingModule,
-        TranslateModule.forRoot({
-            loader: {
-                provide: TranslateLoader,
-                useClass: TranslateLanguageLoader
-            }
-        }),
-        NgxDatatableModule,
-        ToastyModule.forRoot(),
-        TooltipModule.forRoot(),
-        PopoverModule.forRoot(),
-        BsDropdownModule.forRoot(),
-        CarouselModule.forRoot(),
-        ModalModule.forRoot(),
-        ChartsModule
-    ],
-    declarations: [
-        AppComponent,
-        LoginComponent,
-        HomeComponent,
-        CustomersComponent,
-        ProductsComponent,
-        OrdersComponent,
-        SettingsComponent,
-        UsersManagementComponent, UserInfoComponent, UserPreferencesComponent,
-        RolesManagementComponent, RoleEditorComponent,
-        AboutComponent,
-        NotFoundComponent,
-        NotificationsViewerComponent,
-        SearchBoxComponent,
-        StatisticsDemoComponent, TodoDemoComponent, BannerDemoComponent,
-        EqualValidator,
-        LastElementDirective,
-        AutofocusDirective,
-        BootstrapTabDirective,
-        BootstrapToggleDirective,
-        BootstrapSelectDirective,
-        BootstrapDatepickerDirective,
-        GroupByPipe
-    ],
-    providers: [
-        { provide: 'BASE_URL', useFactory: getBaseUrl },
-        { provide: ErrorHandler, useClass: AppErrorHandler },
-        AlertService,
-        ConfigurationService,
-        AppTitleService,
-        AppTranslationService,
-        NotificationService,
-        NotificationEndpoint,
-        AccountService,
-        AccountEndpoint,
-        LocalStoreManager,
-        EndpointFactory
-    ],
-    bootstrap: [AppComponent]
+  imports: [
+    BrowserModule,
+    BrowserAnimationsModule,
+    HttpClientModule,
+    FormsModule,
+    AppRoutingModule,
+    TranslateModule.forRoot({
+      loader: {
+        provide: TranslateLoader,
+        useClass: TranslateLanguageLoader
+      }
+    }),
+    NgxDatatableModule,
+    ToastaModule.forRoot(),
+    TooltipModule.forRoot(),
+    PopoverModule.forRoot(),
+    BsDropdownModule.forRoot(),
+    CarouselModule.forRoot(),
+    ModalModule.forRoot(),
+    ChartsModule
+  ],
+  declarations: [
+    AppComponent,
+    LoginComponent,
+    HomeComponent,
+    CustomersComponent,
+    ProductsComponent,
+    OrdersComponent,
+    SettingsComponent,
+    UsersManagementComponent, UserInfoComponent, UserPreferencesComponent,
+    RolesManagementComponent, RoleEditorComponent,
+    AboutComponent,
+    NotFoundComponent,
+    NotificationsViewerComponent,
+    SearchBoxComponent,
+    StatisticsDemoComponent, TodoDemoComponent, BannerDemoComponent,
+    EqualValidator,
+    LastElementDirective,
+    AutofocusDirective,
+    BootstrapTabDirective,
+    BootstrapToggleDirective,
+    BootstrapSelectDirective,
+    BootstrapDatepickerDirective,
+    GroupByPipe
+  ],
+  providers: [
+    { provide: 'BASE_URL', useFactory: getBaseUrl },
+    { provide: ErrorHandler, useClass: AppErrorHandler },
+    AlertService,
+    ConfigurationService,
+    AppTitleService,
+    AppTranslationService,
+    NotificationService,
+    NotificationEndpoint,
+    AccountService,
+    AccountEndpoint,
+    LocalStoreManager,
+    EndpointFactory
+  ],
+  bootstrap: [AppComponent]
 })
 export class AppModule {
 }
@@ -135,5 +135,5 @@ export class AppModule {
 
 
 export function getBaseUrl() {
-    return document.getElementsByTagName('base')[0].href;
+  return document.getElementsByTagName('base')[0].href;
 }
