@@ -3,16 +3,16 @@
 // Email: support@ebenmonney.com
 // ====================================================
 
-import { Directive, Renderer, ElementRef, OnInit } from '@angular/core';
+import { Directive, ElementRef, OnInit } from '@angular/core';
 
 
 @Directive({
     selector: '[autofocus]'
 })
 export class AutofocusDirective implements OnInit {
-    constructor(public renderer: Renderer, public elementRef: ElementRef) { }
+    constructor(public elementRef: ElementRef) { }
 
     ngOnInit() {
-        setTimeout(() => this.renderer.invokeElementMethod(this.elementRef.nativeElement, 'focus', []), 500);
+        setTimeout(() => this.elementRef.nativeElement['focus'](), 500);
     }
 }
